[nextest-runner] move fixture data into a common crate
We want to reuse this information in the newer integration-tests crate.
