+++ Cargo.toml
@@ -2,22 +2,28 @@
 resolver = "2"
 members = [
   "cargo-nextest",
+  "fixture-data",
   "integration-tests",
+  "nextest-filtering",
   "nextest-metadata",
   "nextest-runner",
-  "nextest-filtering",
   "workspace-hack",
 ]
 
 [workspace.package]
 rust-version = "1.75"
+edition = "2021"
+license = "MIT OR Apache-2.0"
 
 [workspace.dependencies]
 bstr = { version = "1.10.0", default-features = false, features = ["std"] }
 debug-ignore = "1.0.5"
+fixture-data = { path = "fixture-data" }
 globset = "0.4.14"
 guppy = "0.17.7"
+maplit = "1.0.2"
 miette = "7.2.0"
+once_cell = "1.19.0"
 owo-colors = { version = "4.0.0", features = ["supports-colors"] }
 newtype-uuid = { version = "1.1.0", features = ["v4"] }
 nextest-metadata = { version = "0.11.0", path = "nextest-metadata" }
