+++ fixture-data/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "fixture-data"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[dependencies]
+maplit.workspace = true
+nextest-metadata.workspace = true
+nextest-workspace-hack.workspace = true
+once_cell.workspace = true
