+++ fixture-data/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright (c) The nextest Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Information about nextest's fixtures, as contained in the `fixtures` directory.
+
+pub mod models;
+pub mod nextest_tests;
