+++ nextest-runner/Cargo.toml
@@ -133,6 +133,7 @@ self_update = { version = "0.39.0", optional = true }
 
 [dev-dependencies]
 color-eyre = { version = "0.6.3", default-features = false }
+fixture-data.workspace = true
 indoc = "2.0.5"
 insta = { version = "1.39.0", default-features = false }
 maplit = "1.0.2"
