+++ nextest-runner/tests/integration/fixtures.rs
@@ -4,9 +4,10 @@
 use camino::{Utf8Path, Utf8PathBuf};
 use color_eyre::eyre::{Context, Result};
 use duct::cmd;
+use fixture_data::models::FixtureStatus;
 use guppy::{graph::PackageGraph, MetadataCommand};
-use maplit::{btreemap, btreeset};
-use nextest_metadata::{FilterMatch, MismatchReason, RustBinaryId};
+use maplit::btreeset;
+use nextest_metadata::{MismatchReason, RustBinaryId};
 use nextest_runner::{
     cargo_config::{CargoConfigs, EnvironmentMap},
     config::{get_num_cpus, ConfigExperimental, NextestConfig},
@@ -33,176 +34,48 @@ use std::{
     sync::Arc,
 };
 
-#[derive(Copy, Clone, Debug)]
-pub(crate) struct TestFixture {
-    pub(crate) name: &'static str,
-    pub(crate) status: FixtureStatus,
-}
-
-impl PartialEq<(&str, FilterMatch)> for TestFixture {
-    fn eq(&self, (name, filter_match): &(&str, FilterMatch)) -> bool {
-        &self.name == name && self.status.is_ignored() != filter_match.is_match()
-    }
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub(crate) enum FixtureStatus {
-    Pass,
-    Fail,
-    Flaky { pass_attempt: usize },
-    Leak,
-    Segfault,
-    IgnoredPass,
-    IgnoredFail,
-}
-
-impl FixtureStatus {
-    pub(crate) fn to_test_status(self, total_attempts: usize) -> ExecutionResult {
-        match self {
-            FixtureStatus::Pass | FixtureStatus::IgnoredPass => ExecutionResult::Pass,
-            FixtureStatus::Flaky { pass_attempt } => {
-                if pass_attempt <= total_attempts {
-                    ExecutionResult::Pass
-                } else {
-                    ExecutionResult::Fail {
-                        abort_status: None,
-                        leaked: false,
-                    }
-                }
-            }
-            FixtureStatus::Segfault => {
-                cfg_if::cfg_if! {
-                    if #[cfg(unix)] {
-                        // SIGSEGV is 11.
-                        let abort_status = Some(AbortStatus::UnixSignal(11));
-                    } else if #[cfg(windows)] {
-                        // A segfault is an access violation on Windows.
-                        let abort_status = Some(AbortStatus::WindowsNtStatus(
-                            windows_sys::Win32::Foundation::STATUS_ACCESS_VIOLATION,
-                        ));
-                    } else {
-                        let abort_status = None;
-                    }
-                }
+pub(crate) fn make_execution_result(
+    status: FixtureStatus,
+    total_attempts: usize,
+) -> ExecutionResult {
+    match status {
+        FixtureStatus::Pass | FixtureStatus::IgnoredPass => ExecutionResult::Pass,
+        FixtureStatus::Flaky { pass_attempt } => {
+            if pass_attempt <= total_attempts {
+                ExecutionResult::Pass
+            } else {
                 ExecutionResult::Fail {
-                    abort_status,
+                    abort_status: None,
                     leaked: false,
                 }
             }
-            FixtureStatus::Fail | FixtureStatus::IgnoredFail => ExecutionResult::Fail {
-                abort_status: None,
+        }
+        FixtureStatus::Segfault => {
+            cfg_if::cfg_if! {
+                if #[cfg(unix)] {
+                    // SIGSEGV is 11.
+                    let abort_status = Some(AbortStatus::UnixSignal(11));
+                } else if #[cfg(windows)] {
+                    // A segfault is an access violation on Windows.
+                    let abort_status = Some(AbortStatus::WindowsNtStatus(
+                        windows_sys::Win32::Foundation::STATUS_ACCESS_VIOLATION,
+                    ));
+                } else {
+                    let abort_status = None;
+                }
+            }
+            ExecutionResult::Fail {
+                abort_status,
                 leaked: false,
-            },
-            FixtureStatus::Leak => ExecutionResult::Leak,
+            }
         }
-    }
-
-    pub(crate) fn is_ignored(self) -> bool {
-        matches!(
-            self,
-            FixtureStatus::IgnoredPass | FixtureStatus::IgnoredFail
-        )
+        FixtureStatus::Fail | FixtureStatus::IgnoredFail => ExecutionResult::Fail {
+            abort_status: None,
+            leaked: false,
+        },
+        FixtureStatus::Leak => ExecutionResult::Leak,
     }
 }
-
-pub(crate) static EXPECTED_TESTS: Lazy<BTreeMap<RustBinaryId, Vec<TestFixture>>> = Lazy::new(
-    || {
-        btreemap! {
-            // Integration tests
-            "nextest-tests::basic".into() => vec![
-                TestFixture { name: "test_cargo_env_vars", status: FixtureStatus::Pass },
-                TestFixture { name: "test_cwd", status: FixtureStatus::Pass },
-                TestFixture { name: "test_execute_bin", status: FixtureStatus::Pass },
-                TestFixture { name: "test_failure_assert", status: FixtureStatus::Fail },
-                TestFixture { name: "test_failure_error", status: FixtureStatus::Fail },
-                TestFixture { name: "test_failure_should_panic", status: FixtureStatus::Fail },
-                TestFixture { name: "test_flaky_mod_4", status: FixtureStatus::Flaky { pass_attempt: 4 } },
-                TestFixture { name: "test_flaky_mod_6", status: FixtureStatus::Flaky { pass_attempt: 6 } },
-                TestFixture { name: "test_ignored", status: FixtureStatus::IgnoredPass },
-                TestFixture { name: "test_ignored_fail", status: FixtureStatus::IgnoredFail },
-                TestFixture { name: "test_result_failure", status: FixtureStatus::Fail },
-                TestFixture { name: "test_slow_timeout", status: FixtureStatus::IgnoredPass },
-                TestFixture { name: "test_slow_timeout_2", status: FixtureStatus::IgnoredPass },
-                TestFixture { name: "test_slow_timeout_subprocess", status: FixtureStatus::IgnoredPass },
-                TestFixture { name: "test_stdin_closed", status: FixtureStatus::Pass },
-                TestFixture { name: "test_subprocess_doesnt_exit", status: FixtureStatus::Leak },
-                TestFixture { name: "test_success", status: FixtureStatus::Pass },
-                TestFixture { name: "test_success_should_panic", status: FixtureStatus::Pass },
-            ],
-            "nextest-tests::other".into() => vec![
-                TestFixture { name: "other_test_success", status: FixtureStatus::Pass },
-            ],
-            "nextest-tests::segfault".into() => vec![
-                TestFixture { name: "test_segfault", status: FixtureStatus::Segfault },
-            ],
-            // Unit tests
-            "nextest-tests".into() => vec![
-                TestFixture { name: "tests::call_dylib_add_two", status: FixtureStatus::Pass },
-                TestFixture { name: "tests::unit_test_success", status: FixtureStatus::Pass },
-            ],
-            // Binary tests
-            "nextest-tests::bin/nextest-tests".into() => vec![
-                TestFixture { name: "tests::bin_success", status: FixtureStatus::Pass },
-            ],
-            "nextest-tests::bin/other".into() => vec![
-                TestFixture { name: "tests::other_bin_success", status: FixtureStatus::Pass },
-            ],
-            // Example tests
-            "nextest-tests::example/nextest-tests".into() => vec![
-                TestFixture { name: "tests::example_success", status: FixtureStatus::Pass },
-            ],
-            "nextest-tests::example/other".into() => vec![
-                TestFixture { name: "tests::other_example_success", status: FixtureStatus::Pass },
-            ],
-            // Benchmarks
-            "nextest-tests::bench/my-bench".into() => vec![
-                TestFixture { name: "bench_add_two", status: FixtureStatus::Pass },
-                TestFixture { name: "tests::test_execute_bin", status: FixtureStatus::Pass },
-            ],
-            // Proc-macro tests
-            "nextest-derive".into() => vec![
-                TestFixture { name: "it_works", status: FixtureStatus::Pass },
-            ],
-            // Dynamic library tests
-            "cdylib-link".into() => vec![
-                TestFixture { name: "test_multiply_two", status: FixtureStatus::Pass },
-            ],
-            "dylib-test".into() => vec![],
-            "cdylib-example".into() => vec![
-                TestFixture { name: "tests::test_multiply_two_cdylib", status: FixtureStatus::Pass },
-            ],
-            // Build script tests
-            "with-build-script".into() => vec![
-                TestFixture { name: "tests::test_out_dir_present", status: FixtureStatus::Pass },
-            ],
-            "proc-macro-test".into() => vec![],
-        }
-    },
-);
-
-pub(crate) fn get_expected_test(binary_id: &RustBinaryId, test_name: &str) -> &'static TestFixture {
-    let v = EXPECTED_TESTS
-        .get(binary_id)
-        .unwrap_or_else(|| panic!("binary id {binary_id} not found"));
-    v.iter()
-        .find(|fixture| fixture.name == test_name)
-        .unwrap_or_else(|| panic!("for binary id {binary_id}, test name {test_name} not found"))
-}
-
-pub(crate) static EXPECTED_BINARY_LIST: [(&str, &str, bool); 8] = [
-    ("nextest-derive", "nextest-derive", false),
-    ("nextest-tests", "nextest-tests", true),
-    ("nextest-tests::basic", "basic", true),
-    ("nextest-tests::bin/nextest-tests", "nextest-tests", true),
-    ("nextest-tests::bin/other", "other", true),
-    (
-        "nextest-tests::example/nextest-tests",
-        "nextest-tests",
-        true,
-    ),
-    ("nextest-tests::example/other", "other", true),
-    ("nextest-tests::other", "other", true),
-];
 
 #[track_caller]
 pub(crate) fn set_env_vars() {
